requirements.txt
Django==3.1
django-cors-headers==3.2.0
django-filter==21.1
django-rest-auth==0.9.5
djangorestframework==3.13.1
drf-yasg2==1.19.4
mysqlclient==2.1.0
PyMySQL==1.0.2
pytz==2022.1
requests==2.27.1
python-decouple==3.6
cryptography==37.0.4